finish all controller add fuzzy search, init.sql for search, fuzzy dependency, config in application.yaml

+++ src/main/java/com/sarkhan/backend/service/impl/product/ProductServiceImpl.java
@@ -3,15 +3,21 @@ package com.sarkhan.backend.service.impl.product;
 import com.sarkhan.backend.dto.cloudinary.CloudinaryUploadResponse;
 import com.sarkhan.backend.dto.product.ProductFilterRequest;
 import com.sarkhan.backend.dto.product.ProductRequest;
+import com.sarkhan.backend.dto.product.ProductResponseForGetAll;
+import com.sarkhan.backend.dto.product.ProductResponseForSelectedSubCategory;
 import com.sarkhan.backend.mapper.product.ProductMapper;
 import com.sarkhan.backend.model.enums.Role;
 import com.sarkhan.backend.model.product.Product;
+import com.sarkhan.backend.model.product.items.Category;
 import com.sarkhan.backend.model.product.items.Color;
+import com.sarkhan.backend.model.product.items.SubCategory;
 import com.sarkhan.backend.model.user.User;
 import com.sarkhan.backend.repository.product.ProductRepository;
-import com.sarkhan.backend.repository.user.UserRepository;
 import com.sarkhan.backend.service.CloudinaryService;
+import com.sarkhan.backend.service.UserService;
 import com.sarkhan.backend.service.product.ProductService;
+import com.sarkhan.backend.service.product.items.CategoryService;
+import com.sarkhan.backend.service.product.items.SubCategoryService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.access.AccessDeniedException;
@@ -27,8 +33,14 @@ import java.util.*;
 @RequiredArgsConstructor
 public class ProductServiceImpl implements ProductService {
     private final ProductRepository productRepository;
+
     private final CloudinaryService cloudinaryService;
-    private final UserRepository userRepository;
+
+    private final CategoryService categoryService;
+
+    private final SubCategoryService subCategoryService;
+
+    private final UserService userService;
 
     @Override
     public Product add(ProductRequest request, List<MultipartFile> images) throws IOException {
@@ -47,9 +59,12 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> getAll() {
+    public ProductResponseForGetAll getAll() {
         log.info("Someone try to get all products.");
-        return productRepository.findAll();
+        return new ProductResponseForGetAll(
+                productRepository.findAll(),
+                categoryService.getAll(),
+                subCategoryService.getAll());
     }
 
     @Override
@@ -71,33 +86,61 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public List<Product> searchByName(String name) {
+    public ProductResponseForGetAll searchByName(String name) {
         log.info("Someone try to get products. Name : " + name);
-        return productRepository.searchByName(name);
+        List<Product> products = productRepository.searchByName(name);
+        List<Category> categories = categoryService.searchByName(name);
+        Set<SubCategory> subCategories = new HashSet<>();
+
+        for (Category category : categories)
+            subCategories.addAll(subCategoryService.getByCategoryId(category.getId()));
+
+        subCategories.addAll(subCategoryService.searchByName(name));
+
+        for (SubCategory subCategory : subCategories)
+            products.addAll(productRepository.getBySubCategoryId(subCategory.getId()));
+
+        return new ProductResponseForGetAll(
+                products,
+                categoryService.getAll(),
+                subCategoryService.getAll());
     }
 
     @Override
-    public List<Product> getBySubCategoryId(Long subCategoryId) {
+    public ProductResponseForSelectedSubCategory getBySubCategoryId(Long subCategoryId) {
         log.info("Someone try to get products. SubCategory id : " + subCategoryId);
-        return productRepository.getBySubCategoryId(subCategoryId);
+        return new ProductResponseForSelectedSubCategory(
+                productRepository.getBySubCategoryId(subCategoryId),
+                categoryService.getAll(),
+                subCategoryService.getAll(),
+                subCategoryService.getById(subCategoryId).getSpecifications(),
+                new ProductFilterRequest(subCategoryId, null, null, null, null, null));
     }
 
     @Override
-    public List<Product> getBySellerId(Long sellerId) {
+    public ProductResponseForGetAll getBySellerId(Long sellerId) {
         log.info("Someone try to get products. Seller id : " + sellerId);
-        return productRepository.getBySellerId(sellerId);
+        return new ProductResponseForGetAll(
+                productRepository.getBySellerId(sellerId),
+                categoryService.getAll(),
+                subCategoryService.getAll());
     }
 
     @Override
-    public List<Product> getByComplexFiltering(ProductFilterRequest request) {
+    public ProductResponseForSelectedSubCategory getByComplexFiltering(ProductFilterRequest request) {
         log.info("Someone try to get product with complex params.");
-        return productRepository.getByComplexFiltering(request);
+        return new ProductResponseForSelectedSubCategory(
+                productRepository.getByComplexFiltering(request),
+                categoryService.getAll(),
+                subCategoryService.getAll(),
+                subCategoryService.getById(request.subCategoryId()).getSpecifications(),
+                request);
     }
 
     @Override
-    public Product giveRating(Long productId, Double rating) {
+    public Product giveRating(Long id, Double rating) {
         User user = getCurrentUser();
-        Product product = getById(productId);
+        Product product = getById(id);
 
         log.info(user.getNameAndSurname() + " try to give rating. Product name : " + product.getName());
 
@@ -116,9 +159,9 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product toggleFavorite(Long productId) {
+    public Product toggleFavorite(Long id) {
         User user = getCurrentUser();
-        Product product = getById(productId);
+        Product product = getById(id);
 
         log.info(user.getNameAndSurname() + " pres favorite button. Product name : " + product.getName());
 
@@ -140,7 +183,10 @@ public class ProductServiceImpl implements ProductService {
         Product oldProduct = ProductMapper.updateEntity(getById(id), request);
         User user = getCurrentUser();
 
-        if (!oldProduct.getSellerId().equals(user.getId()) || Role.ADMIN.equals(user.getRole())) {
+        log.info(user.getNameAndSurname() + " try to update product. Id : " + id);
+
+        if (!(Role.ADMIN.equals(user.getRole()) || getById(id).getSellerId().equals(user.getId()))) {
+            log.warn("User is not authorized to update this product");
             throw new AccessDeniedException("You are not authorized to update this product");
         }
 
@@ -160,15 +206,16 @@ public class ProductServiceImpl implements ProductService {
     public void delete(Long id) {
         User user = getCurrentUser();
         log.warn(user.getNameAndSurname() + " delete product. Id : " + id);
-        productRepository.deleteById(id);
+        if (Role.ADMIN.equals(user.getRole()) || getById(id).getSellerId().equals(user.getId())) {
+            productRepository.deleteById(id);
+        } else {
+            log.warn(user.getNameAndSurname() + " cannot delete this product.");
+        }
     }
 
     private User getCurrentUser() {
         String email = (String) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        return userRepository.findByEmail(email).orElseThrow(() -> {
-            log.warn("Cannot find user by email: " + email);
-            return new NoSuchElementException("Cannot find user by email: " + email);
-        });
+        return userService.getByEmail(email);
     }
 
     private List<Color> uploadImages(ProductRequest request, List<MultipartFile> images) throws IOException {
